Create Deck Shuffler
This was a project done while taking a Swift programming course in 2015, so this is using older Swift syntax.

It's a deck shuffler app. I didn't get to finish this because I had found work, but I was last working on creating a second array to store the cards that had been drawn, so only the cards that were remaining could be drawn.

I was learning this in between jobs, and had a mentor who, despite being unfamiliar with Swift, was guiding me by teaching me how to think about structuring my code. The code which generates cards in a deck was actually his suggestion, but I wrote it.