Create HowManyFingers.swift
This was a project in the Swift class I was taking. It was a game that allows the app to generate a number of fingers and let the player guess how many fingers the app had.

Pretty simple, but the most confusing thing for me here was having all these different types - Int, UInt32, etc. That part is probably the most unintuitive part for a newbie.