Create HowManyFingers2
This is an updated version of the previous HowManyFingers app I made. In the previous version, to guess the number of fingers, there was an input box which brought up the keyboard.

The UX was really annoying for me, so I decided to create buttons representing answers, instead. This also limited the amount of tooltips needed, because now I didn't need to warn the users of the limit of 1-5 fingers only. The UI would speak for itself.

What this taught me is that good UX can mean more complicated code for the developers, which is probably why my teammates in game development got slightly annoyed whenever I'd suggest improvements to app UI flows.